esc, put hook funcs in new file
