+++ tutorial/circle.c
@@ -40,16 +40,6 @@ bool is_crossed(double x_img, double y_img, t_vecs *vecs)
 		return (false);
 }
 
-int ft_exit(t_data *data)
-{
-	mlx_destroy_image(data->mlx, data->img);
-	mlx_destroy_window(data->mlx, data->mlx_win);
-	mlx_destroy_display(data->mlx);
-	free(data->mlx);
-
-	exit(0);
-}
-
 int	main(void)
 {
 	t_data	data;
@@ -85,6 +75,7 @@ int	main(void)
 	}
 
 	mlx_hook(data.mlx_win, 33, 1 << 17, ft_exit, &data);
+	mlx_key_hook(data.mlx_win, key_hook, &data);
 	mlx_put_image_to_window(data.mlx, data.mlx_win, data.img, 0, 0);
 	mlx_loop(data.mlx);
 }
